full page snapshot if issue with content selector + logic about mail in case of errors
