+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 2021-06-04
+*Fixes*
+- implemented logic about sending mail with the info about error (and screenshot of the full page) in case of issues with selectors during the actor run;
+- Implemented snapshot of the whole page if content selector is not found (issue #18).
+
 ## 2021-06-03
 *Fixes*
 - Updated SDK to 1.2.0;
